LPD-84340 Move SectionHeader to shared components
