+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/individual/profile/components/AccountMembership.tsx
@@ -10,7 +10,7 @@ import {
 } from 'shared/components/GeneralInfoSection';
 import {formatUTCDate} from 'shared/util/date';
 import {Map} from 'immutable';
-import {SectionHeader} from './SectionHeader';
+import {SectionHeader} from 'shared/components/SectionHeader';
 
 function formatCurrency(currencyCode: string, value: string): string {
 	return new Intl.NumberFormat(undefined, {
