+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/individual/profile/components/IndividualAttributesCDP.tsx
@@ -4,7 +4,7 @@ import {
 	GeneralInfoSection
 } from 'shared/components/GeneralInfoSection';
 import {formatUTCDate} from 'shared/util/date';
-import {SectionHeader} from './SectionHeader';
+import {SectionHeader} from 'shared/components/SectionHeader';
 
 interface IIndividualAttributesProps {
 	children?: React.ReactNode;
