+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/shared/components/SectionHeader.tsx
